Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -4,8 +4,8 @@ int main(){
     float pi = 3.14;
     float area;
     int r;
-    scanf("%f", &r);
-    area = pi * r * r;
+    scanf("%d", &r);
+    area = (float)pi * r * r;
     printf("Area: %.2f", area);
 
     return 0;
